Bug fix for 2 most recent releases: s/OSError/EnvironmentError/g
This is a follow up to / bug fix for 2163f771e1 and b05880cf8e where I
accidentally used OSError when I should have used IOError or the generic
parent class EnvironmentError. I've verified that EnvironmentError is
compatible with both Python 2.7 and 3.

The last few releases clearly show the value of having an automated test
suite to guard against regressions, unfortunately due to the nature of
the negotiator project I'm not quite sure how I would get a functional
test suite up and running on Travis CI.

It's definitely possible, and on my wish list, but it is one of the
higher hanging fruits on that -almost infinite- wish list, so don't hold
your breath 😛.